requirements.txt
python-telegram-bot==20.6
psycopg2-binary
feedparser
spongemock
zalgo-text
regex
bleach
telethon
alphabet_detector
pyrate-limiter
cachetools
numpy
pymongo
httpcore
httpx
python-telegram-bot[job-queue]
web3
pytz